Added argument inclusion and exclusion list
SimpleArg now has an InclusionList and ExclusionList property. It's a List<string> and its valus is optional.

InclusionList: If this argument is passed then all named arguments in the InclusionList must be passed.

ExclusionList: If this argument is passed then no named argument in the ExclusionList can be passed.

Only InclusionList and ExclusionsList properties in Required arguments are checked. If optional arguments specify Inclusion or ExclusionLists they will be ignored.

Only optional arguments can be named in Inclusion and ExclusionList.